backend/requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
openpyxl>=3.1.0,<4.0.0

# Visualization
matplotlib>=3.7.0,<4.0.0
plotly>=5.0.0,<6.0.0

# PDF Generation
reportlab>=4.0.0,<5.0.0

# AI & LLM
langchain>=0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0

# Vector Database & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Advanced Analytics
scikit-learn>=1.3.0
prophet>=1.1.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0

# CORS
fastapi-cors>=0.0.6
